Added check if entity exists to soncontroller delete

+++ src/main/java/de/dhbw/ti21/webeng2/streaming_playlist/controller/SongController.java
@@ -81,6 +81,10 @@ public class SongController {
     @DeleteMapping
     public ResponseEntity<Void> deleteSong(@RequestParam Long id){
         try{
+            if (!this.songRepository.existsById(id)){
+                return ResponseEntity.noContent().build();
+            }
+
             this.songRepository.deleteById(id);
             return ResponseEntity.ok().build();
         }
